fix: adjust linting
## Rust linter (clippy)

A recent update to rust's `clippy` tool triggered several lint violations when not targeting Linux.

Some just required an explicitly anonymous lifetime for member functions that return buffer objects in the bindings.

Others were about unused structs on non-Linux platforms in the node.js binding.
The solution for this one was to conditionally compile the structs for only Linux targets.
This is not considered a breaking change because the type info remains the same regardless.

## Pre-commit hooks

Updated the pre-commit hooks and ran them

## Metadata

Some metadata changes were made to exclude linting the rust examples by default.
This was done because the examples do not compile for all platforms.
More attention is still needed on this subject (see #8).

To run the examples (still only on Linux), an extra arg is now required: `-p examples`.
So, the command to run the getting-started example is now:
```shell
cargo run -p examples --bin getting-started
```

These metadata changes also incited changes to the nurfile about various command invoked in tasks.

## CI workflows

I adjusted the lint CI workflow to check all rust packages (including examples) in the workspace (this repo).

I reviewed the caching of rust assets, which are only used when linting, testing, or documenting rust libraries.